fix problems in 1015cd4ff73ee25e55b4da8c8bc55a41e652da87:
- inverted check in SwDoc::RemoveInvisibleContent
- various too short end positions in DeleteAndDestroy calls in
  SwDoc::~SwDoc and SwDoc::ClearDoc
- std::copy to begin() of empty vector in SwReader::Read replaced
  with std::back_inserter; the vector is used as a set here so the
  end position should work just as well

+++ sw/source/core/unocore/unotext.cxx
@@ -1668,7 +1668,7 @@ throw (lang::IllegalArgumentException, uno::RuntimeException)
 
     // see if there are frames already anchored to this node
     std::vector<SwFrmFmt*> aAnchoredFrames;
-    for (int i = 0; i < (int)m_pImpl->m_pDoc->GetSpzFrmFmts()->size(); ++i)
+    for (size_t i = 0; i < m_pImpl->m_pDoc->GetSpzFrmFmts()->size(); ++i)
     {
         SwFrmFmt* pFrmFmt = (*m_pImpl->m_pDoc->GetSpzFrmFmts())[i];
         const SwFmtAnchor& rAnchor = pFrmFmt->GetAnchor();
